Refactor Rust gatekeeper service shell
